README: Add a note about No Longer Maintained
Hi @njouanin,

It's not immediately obvious to people who find this project that development has moved to the aqmtt fork.

This is a small PR to make this information easier to find, please consider merging it. Thanks for all your work on this project, I'm glad to see that it continues to have an active life as aqmtt. :)